requirements.txt
Django==2.0.3
django-bootstrap3==9.1.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1